requirements.txt
requests
python-dotenv
playwright
pytest
pytest-json-report
pytest-html
pytest-playwright
openpyxl
anthropic>=0.18.0